refactor: more simplest architecture
