+++ package.json
@@ -6,10 +6,10 @@
     "url": ""
   },
   "description": "REST service starter for RS School",
-  "main": "dist/server.js",
+  "main": "dist/index.js",
   "scripts": {
-    "start": "node dist/server.js",
-    "dev": "nodemon --config nodemon.json src/server.ts",
+    "start": "node dist/index.js",
+    "dev": "nodemon --config nodemon.json src/index.ts",
     "build": "tsc --build tsconfig.json",
     "lint": "eslint --ext .js,.ts --fix src/",
     "test": "cross-env DEBUG=rs:* jest --testMatch \"<rootDir>/test/e2e/test/*.test.js\" --noStackTrace --runInBand",
